requirements.txt
arxiv
google-generativeai